Updated examples

+++ TORCphysics/src/binding_model.py
@@ -1180,7 +1180,7 @@ class TopoIRecognition(BindingModel):
         a = environmental.concentration * self.k_on
         # b1 = np.float128((superhelical + self.threshold) / self.width)
         #b1 = np.float128((superhelical - self.threshold) / self.width)
-        b1 = np.float((superhelical - self.threshold) / self.width)
+        b1 = (superhelical - self.threshold) / self.width
         b = 1 + np.exp(b1)
         rate = a / b
         return utils.P_binding_Nonh_Poisson(rate=rate, dt=dt)
@@ -1302,7 +1302,7 @@ class TopoIRecognitionRNAPTracking(BindingModel):
         a = environmental.concentration * self.k_on
         # b1 = np.float128((superhelical + self.threshold) / self.width)
         #b1 = np.float128((superhelical - self.threshold) / self.width)
-        b1 = np.float((superhelical - self.threshold) / self.width)
+        b1 = (superhelical - self.threshold) / self.width
         b = 1 + np.exp(b1)
         rate = a / b
         # b = 1 + np.exp((superhelical - self.threshold) / self.width)
@@ -1438,7 +1438,7 @@ class GyraseRecognition(BindingModel):
         a = environmental.concentration * self.k_on
         # b1 = np.float128(-(superhelical + self.threshold) / self.width)
         #b1 = np.float128(-(superhelical - self.threshold) / self.width)
-        b1 = np.float(-(superhelical - self.threshold) / self.width)
+        b1 = -(superhelical - self.threshold) / self.width
         b = 1 + np.exp(b1)
         rate = a / b
         # Configure NumPy to suppress overflow warnings
